utils: factor out Cp2kInit for unittesting
also fix the recursion bug introduced by doing

    isinstance(val, Sequence)

when `val` being a string.
